@@ -1,17 +1,16 @@
-"use client"
+'use client';
 
-import React from 'react'
-import Image from 'next/image'
-import AddFile from './AddFile'
-import FileSelectOptions from './FileSelectOptions'
-import {AnimatePresence, motion, useAnimation} from 'framer-motion'
+import React from 'react';
+import AddFile from './AddFile';
+import FileSelectOptions from './FileSelectOptions';
+import { AnimatePresence, motion, useAnimation } from 'framer-motion';
 import { selectedFilesStore } from '@/utils/store/selectFilesStore';
-import UploadFileModal from '@/components/Modals/UploadFileModal'
+import UploadFileModal from '@/components/Modals/UploadFileModal';
 
 const FileOperations = () => {
   const [addFile, setAddFile] = React.useState(false);
   const [files] = selectedFilesStore((state) => [state.files]);
-  const controls = useAnimation(); 
+  const controls = useAnimation();
 
   const toggleAddFile = () => {
     setAddFile(!addFile);
@@ -19,28 +18,31 @@ const FileOperations = () => {
   };
 
   return (
-    <div className='Operations flex justify-between pr-1'>
-      <div className='flex gap-6'>
-        <UploadFileModal/>
+    <div className="Operations flex justify-between pr-1">
+      <div className="flex gap-6">
+        <UploadFileModal />
 
-        <motion.button 
+        <motion.button
           className="bg-primary_bg px-[0.65rem] pb-[0.4rem] rounded-full hover:bg-button_hover"
-          onClick={toggleAddFile} 
-          animate={controls} 
+          onClick={toggleAddFile}
+          animate={controls}
         >
-          <p title='Add File/Folder' className='leading-[0px] text-3xl text-secondary_font font'>+</p>
+          <p
+            title="Add File/Folder"
+            className="leading-[0px] text-3xl text-secondary_font font"
+          >
+            +
+          </p>
         </motion.button>
 
-        <AnimatePresence>
-          {addFile && <AddFile />}
-        </AnimatePresence>
+        <AnimatePresence>{addFile && <AddFile />}</AnimatePresence>
       </div>
 
       <AnimatePresence>
-        {files.length > 0 && <FileSelectOptions/>}
+        {files.length > 0 && <FileSelectOptions />}
       </AnimatePresence>
     </div>
-  )
-}
+  );
+};
 
-export default FileOperations
+export default FileOperations;
