@@ -1,20 +1,20 @@
-"use client"
+'use client';
 
-import React, {useState} from 'react'
+import React, { useState } from 'react';
 import {
-    AlertDialog,
-    AlertDialogAction,
-    AlertDialogCancel,
-    AlertDialogContent,
-    AlertDialogDescription,
-    AlertDialogFooter,
-    AlertDialogHeader,
-    AlertDialogTitle,
-    AlertDialogTrigger,
-  } from "@/components/ui/alert-dialog"
-import Image from 'next/image'
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from '@/components/ui/alert-dialog';
+import Image from 'next/image';
 import { motion, AnimatePresence } from 'framer-motion';
-import SelectedFilesDisplay from '@/components/Modals/Modal-components/SelectedFilesDisplay'
+import SelectedFilesDisplay from '@/components/Modals/Modal-components/SelectedFilesDisplay';
 import { selectedFilesStore } from '@/utils/store/selectFilesStore';
 import TabSelectionComponent from './TabSelectionComponent';
 import GenerateLinkSection from './GenerateLinkSection';
@@ -28,15 +28,20 @@ import groupShare from '@/utils/api/shareInGroupAPI';
 
 type Props = {
   multiplefiles: boolean;
-}
+};
 
-const ShareContentModal = ({multiplefiles}:Props) => {
-  const [tab, setTab] = useState<"link" | "email" | "internal" | "groups">('link');
-  const [files, removeFile, removeAllFiles] = selectedFilesStore((state) => [state.files, state.removeFile, state.removeAllFiles]);
+const ShareContentModal = ({ multiplefiles }: Props) => {
+  const [tab, setTab] = useState<'link' | 'email' | 'internal' | 'groups'>(
+    'link'
+  );
+  const [files, removeFile] = selectedFilesStore((state) => [
+    state.files,
+    state.removeFile,
+  ]);
   const [linkName, setLinkName] = useState<string>('');
   const [shareEmail, setShareEmail] = useState<string>('');
   const [sharedSuccessfully, setSharedSuccessfully] = useState<boolean>(false);
-  const [response, setResponse] = useState<any>(null);
+  const [response, setResponse] = useState(null);
   const [selectedGroups, setSelectedGroups] = useState<string[]>([]);
   const [allChecked, setAllChecked] = React.useState<boolean>(false);
 
@@ -50,15 +55,16 @@ const ShareContentModal = ({multiplefiles}:Props) => {
     expiryDate: null,
     expiryTime: null,
     passwordValue: null,
-    accessValue: 0
+    accessValue: 0,
   });
 
-  const [shareInternalSettings, setShareInternalSettings] = useState<InternalShareSettings>({
-    shareable: false,
-    downloadable: false,
-    proctored: false,
-    modifyable: false,
-  });
+  const [shareInternalSettings, setShareInternalSettings] =
+    useState<InternalShareSettings>({
+      shareable: false,
+      downloadable: false,
+      proctored: false,
+      modifyable: false,
+    });
 
   const removeFileFromSelection = (indexToRemove: number) => {
     removeFile(files[indexToRemove].urlhash);
@@ -67,16 +73,16 @@ const ShareContentModal = ({multiplefiles}:Props) => {
   const onCloseAlert = () => {
     setSharedSuccessfully(false);
     setResponse(null);
-  }
+  };
 
   const handleShare = async () => {
     try {
-      const file = files.filter(item => item.is_file);
-      const folder = files.filter(item => !item.is_file);
-  
-      const fileUrls = file.map(item => item.urlhash);
-      const folderUrls = folder.map(item => item.urlhash);
-  
+      const file = files.filter((item) => item.is_file);
+      const folder = files.filter((item) => !item.is_file);
+
+      const fileUrls = file.map((item) => item.urlhash);
+      const folderUrls = folder.map((item) => item.urlhash);
+
       let shareResponse;
       if (tab === 'internal') {
         shareResponse = await internalShare(
@@ -120,7 +126,7 @@ const ShareContentModal = ({multiplefiles}:Props) => {
         expiryDate: null,
         expiryTime: null,
         passwordValue: null,
-        accessValue: 0
+        accessValue: 0,
       });
       setSelectedGroups([]);
       setLinkName('');
@@ -132,36 +138,62 @@ const ShareContentModal = ({multiplefiles}:Props) => {
     } catch (error) {
       console.error(error);
     }
-  };  
+  };
 
   const variants = {
     open: { opacity: 1, scale: 1 },
     closed: { opacity: 0, scale: 0.9 },
-  };  
+  };
 
   return (
     <AlertDialog>
       <AlertDialogTrigger asChild>
-        {multiplefiles 
-          ? <Image src='/share-icon.svg' onClick={(e) => e.stopPropagation()} width={20} height={20} alt='Share icon'/>
-          : <p onClick={(e) => e.stopPropagation()} className='w-full'>Share</p>
-        }
+        {multiplefiles ? (
+          <Image
+            src="/share-icon.svg"
+            onClick={(e) => e.stopPropagation()}
+            width={20}
+            height={20}
+            alt="Share icon"
+          />
+        ) : (
+          <p onClick={(e) => e.stopPropagation()} className="w-full">
+            Share
+          </p>
+        )}
       </AlertDialogTrigger>
-      <AlertDialogContent className='w-[50rem] gap-1'>
-        <AlertDialogHeader className='flex flex-row h-10 justify-between'>
-          <AlertDialogTitle className='flex gap-4 items-center pt-1 '>
-            <Image src='/share-icon.svg' className='rounded-full p-[0.35rem] bg-primary_bg' width={30} height={30} alt='Share icon'/>
+      <AlertDialogContent className="w-[50rem] gap-1">
+        <AlertDialogHeader className="flex flex-row h-10 justify-between">
+          <AlertDialogTitle className="flex gap-4 items-center pt-1 ">
+            <Image
+              src="/share-icon.svg"
+              className="rounded-full p-[0.35rem] bg-primary_bg"
+              width={30}
+              height={30}
+              alt="Share icon"
+            />
             <p>Share Content</p>
           </AlertDialogTitle>
-          <AlertDialogCancel className='w-7 h-7 p-[0.4rem] rounded-full bg-[#F0F0F0] mt-0' onClick={(e) => e.stopPropagation()}>
-            <Image src="/cross-icon.svg" width={20} height={20} className='rounded-full' alt='close icon'/>
+          <AlertDialogCancel
+            className="w-7 h-7 p-[0.4rem] rounded-full bg-[#F0F0F0] mt-0"
+            onClick={(e) => e.stopPropagation()}
+          >
+            <Image
+              src="/cross-icon.svg"
+              width={20}
+              height={20}
+              className="rounded-full"
+              alt="close icon"
+            />
           </AlertDialogCancel>
         </AlertDialogHeader>
-        <AlertDialogDescription className='h-[26rem] text-md p-2'>
+        <AlertDialogDescription className="h-[26rem] text-md p-2">
+          <SelectedFilesDisplay
+            file={files}
+            removeFileFromSelection={removeFileFromSelection}
+          />
 
-          <SelectedFilesDisplay file={files} removeFileFromSelection={removeFileFromSelection}/>
-
-          <TabSelectionComponent tab={tab} setTab={setTab}/>
+          <TabSelectionComponent tab={tab} setTab={setTab} />
 
           <AnimatePresence initial={false}>
             <motion.div
@@ -172,65 +204,89 @@ const ShareContentModal = ({multiplefiles}:Props) => {
               transition={{
                 type: 'spring',
                 stiffness: 200,
-                damping: 20
+                damping: 20,
               }}
               key={tab}
-              className='absolute h-[25rem] w-[46rem]'
+              className="absolute h-[25rem] w-[46rem]"
             >
-          {tab === "link" && (
-            <GenerateLinkSection settings={shareSettings} setSettings={setShareSettings} linkName={linkName} setLinkName={setLinkName}/>
-          )}
+              {tab === 'link' && (
+                <GenerateLinkSection
+                  settings={shareSettings}
+                  setSettings={setShareSettings}
+                  linkName={linkName}
+                  setLinkName={setLinkName}
+                />
+              )}
 
-          {tab === "email" && (
-            <SendMailSection settings={shareSettings} setSettings={setShareSettings} linkName={linkName} setLinkName={setLinkName} shareEmail={shareEmail} setShareEmail={setShareEmail}/>
-          )}
+              {tab === 'email' && (
+                <SendMailSection
+                  settings={shareSettings}
+                  setSettings={setShareSettings}
+                  linkName={linkName}
+                  setLinkName={setLinkName}
+                  shareEmail={shareEmail}
+                  setShareEmail={setShareEmail}
+                />
+              )}
 
-          {tab === "internal" && (
-            <InternalShareSection shareEmail={shareEmail} setShareEmail={setShareEmail} settings={shareInternalSettings} setSettings={setShareInternalSettings}/>
-          )}
+              {tab === 'internal' && (
+                <InternalShareSection
+                  shareEmail={shareEmail}
+                  setShareEmail={setShareEmail}
+                  settings={shareInternalSettings}
+                  setSettings={setShareInternalSettings}
+                />
+              )}
 
-          {tab === "groups" && (
-            <SendInGroupsSection settings={shareInternalSettings} setSettings={setShareInternalSettings} selectedGroups={selectedGroups} setSelectedGroups={setSelectedGroups} allChecked={allChecked} setAllChecked={setAllChecked}/>
-          )}
+              {tab === 'groups' && (
+                <SendInGroupsSection
+                  settings={shareInternalSettings}
+                  setSettings={setShareInternalSettings}
+                  selectedGroups={selectedGroups}
+                  setSelectedGroups={setSelectedGroups}
+                  allChecked={allChecked}
+                  setAllChecked={setAllChecked}
+                />
+              )}
             </motion.div>
           </AnimatePresence>
         </AlertDialogDescription>
-        <AlertDialogFooter className='flex justify-end'>
-          {tab === "link" && (
-            <AlertDialogAction 
-              onClick={handleShare} 
-              className='rounded-full text-white font-normal bg-primary_font_2 hover:text-primary_font_2 border-[1px] border-solid border-primary_font_2'
+        <AlertDialogFooter className="flex justify-end">
+          {tab === 'link' && (
+            <AlertDialogAction
+              onClick={handleShare}
+              className="rounded-full text-white font-normal bg-primary_font_2 hover:text-primary_font_2 border-[1px] border-solid border-primary_font_2"
               disabled={!linkName}
             >
               Generate link
             </AlertDialogAction>
           )}
 
-          {tab === "email" && (
-            <AlertDialogAction 
-              onClick={handleShare} 
-              className='rounded-full text-white font-normal bg-primary_font_2 hover:text-primary_font_2 border-[1px] border-solid border-primary_font_2'
+          {tab === 'email' && (
+            <AlertDialogAction
+              onClick={handleShare}
+              className="rounded-full text-white font-normal bg-primary_font_2 hover:text-primary_font_2 border-[1px] border-solid border-primary_font_2"
               disabled={!linkName || !shareEmail}
             >
               Send Mail
             </AlertDialogAction>
           )}
 
-          {tab === "internal" && (
+          {tab === 'internal' && (
             <AlertDialogAction
               onClick={handleShare}
               disabled={!shareEmail}
-              className='rounded-full text-white font-normal bg-primary_font_2 hover:text-primary_font_2 border-[1px] border-solid border-primary_font_2'
+              className="rounded-full text-white font-normal bg-primary_font_2 hover:text-primary_font_2 border-[1px] border-solid border-primary_font_2"
             >
               Share internally
             </AlertDialogAction>
           )}
 
-          {tab === "groups" && (
+          {tab === 'groups' && (
             <AlertDialogAction
               onClick={handleShare}
               disabled={selectedGroups.length === 0}
-              className='rounded-full text-white font-normal bg-primary_font_2 hover:text-primary_font_2 border-[1px] border-solid border-primary_font_2'
+              className="rounded-full text-white font-normal bg-primary_font_2 hover:text-primary_font_2 border-[1px] border-solid border-primary_font_2"
             >
               Share in groups
             </AlertDialogAction>
@@ -238,38 +294,71 @@ const ShareContentModal = ({multiplefiles}:Props) => {
         </AlertDialogFooter>
       </AlertDialogContent>
 
-      {(sharedSuccessfully && tab==="link") && (
-        <BotLeftAlert image='/task-completed-icon.svg' imagebg='bg-[#E5EDFF]' className='w-fit'>
-          <div className='flex flex-col items-start text-left leading-[0.2rem] gap-[0.35rem]'>
-            <p className='text-primary_font font-semibold text-base leading-4  '>Files/Folders shared successfully</p>
-            <p className='text-[#979797] font-[400] text-sm leading-[1.1rem]' title={response.data.data.link}>Link: {response.data.data.link}</p>
+      {sharedSuccessfully && tab === 'link' && (
+        <BotLeftAlert
+          image="/task-completed-icon.svg"
+          imagebg="bg-[#E5EDFF]"
+          className="w-fit"
+        >
+          <div className="flex flex-col items-start text-left leading-[0.2rem] gap-[0.35rem]">
+            <p className="text-primary_font font-semibold text-base leading-4  ">
+              Files/Folders shared successfully
+            </p>
+            <p
+              className="text-[#979797] font-[400] text-sm leading-[1.1rem]"
+              title={response.data.data.link}
+            >
+              Link: {response.data.data.link}
+            </p>
           </div>
-          <button onClick={onCloseAlert} className='text-primary_font px-2 mb-8 text-lg'>x</button>
+          <button
+            onClick={onCloseAlert}
+            className="text-primary_font px-2 mb-8 text-lg"
+          >
+            x
+          </button>
         </BotLeftAlert>
       )}
 
-      {(sharedSuccessfully && tab==="email") && (
-        <BotLeftAlert image='/task-completed-icon.svg' imagebg='bg-[#E5EDFF]'>
-          <div className='flex flex-col items-start text-left leading-[0.2rem] gap-[0.35rem]'>
-            <p className='text-primary_font font-semibold text-base leading-4  '>Files/Folders shared successfully</p>
-            <p className='text-[#979797] font-[400] text-sm leading-[1.1rem]'>{response.data.message}</p>
+      {sharedSuccessfully && tab === 'email' && (
+        <BotLeftAlert image="/task-completed-icon.svg" imagebg="bg-[#E5EDFF]">
+          <div className="flex flex-col items-start text-left leading-[0.2rem] gap-[0.35rem]">
+            <p className="text-primary_font font-semibold text-base leading-4  ">
+              Files/Folders shared successfully
+            </p>
+            <p className="text-[#979797] font-[400] text-sm leading-[1.1rem]">
+              {response.data.message}
+            </p>
           </div>
-          <button onClick={onCloseAlert} className='text-primary_font px-2 mb-8 text-lg'>x</button>
+          <button
+            onClick={onCloseAlert}
+            className="text-primary_font px-2 mb-8 text-lg"
+          >
+            x
+          </button>
         </BotLeftAlert>
       )}
 
-      {(sharedSuccessfully && tab==="groups") && (
-        <BotLeftAlert image='/task-completed-icon.svg' imagebg='bg-[#E5EDFF]'>
-          <div className='flex flex-col items-start text-left leading-[0.2rem] gap-[0.35rem]'>
-            <p className='text-primary_font font-semibold text-base leading-4  '>Files/Folders shared successfully</p>
-            <p className='text-[#979797] font-[400] text-sm leading-[1.1rem]'>{response.data.message.trial}</p>
+      {sharedSuccessfully && tab === 'groups' && (
+        <BotLeftAlert image="/task-completed-icon.svg" imagebg="bg-[#E5EDFF]">
+          <div className="flex flex-col items-start text-left leading-[0.2rem] gap-[0.35rem]">
+            <p className="text-primary_font font-semibold text-base leading-4  ">
+              Files/Folders shared successfully
+            </p>
+            <p className="text-[#979797] font-[400] text-sm leading-[1.1rem]">
+              {response.data.message.trial}
+            </p>
           </div>
-          <button onClick={onCloseAlert} className='text-primary_font px-2 mb-8 text-lg'>x</button>
+          <button
+            onClick={onCloseAlert}
+            className="text-primary_font px-2 mb-8 text-lg"
+          >
+            x
+          </button>
         </BotLeftAlert>
       )}
-
     </AlertDialog>
-  )
-}
+  );
+};
 
-export default ShareContentModal
+export default ShareContentModal;
