@@ -1,15 +1,13 @@
-"use client"
+'use client';
 
-import React,{useEffect, useState} from 'react'
-import UserLinkAnalytics from '@/components/User-dashboard/UserDashLink'
-import UserDashStorage from './UserDashStorage'
-import UserDashFavGroups from './UserDashFavGroups'
-import UserDashCritLinks from './UserDashCritLinks'
-import getUserDetails from '@/utils/api/getUserDashInfoAPI'
+import { useEffect, useState } from 'react';
+import UserLinkAnalytics from '@/components/User-dashboard/UserDashLink';
+import UserDashStorage from './UserDashStorage';
+import UserDashFavGroups from './UserDashFavGroups';
+import UserDashCritLinks from './UserDashCritLinks';
+import getUserDetails from '@/utils/api/getUserDashInfoAPI';
 
-type Props = {}
-
-const UserDashboard = (props: Props) => {
+const UserDashboard = () => {
   const [userDetails, setUserDetails] = useState<UserDashDetails>();
 
   useEffect(() => {
@@ -17,7 +15,7 @@ const UserDashboard = (props: Props) => {
       try {
         const result = await getUserDetails();
         setUserDetails(result);
-        console.log(result)
+        console.log(result);
       } catch (error) {
         console.error(error);
       }
@@ -29,20 +27,20 @@ const UserDashboard = (props: Props) => {
   }, []);
 
   return (
-    <div className='flex gap-9 flex-col w-full h-full'>
-      <div className='flex gap-10 h-[55%]'>
-        <UserLinkAnalytics linkDetails={userDetails?.links}/>
+    <div className="flex gap-9 flex-col w-full h-full">
+      <div className="flex gap-10 h-[55%]">
+        <UserLinkAnalytics linkDetails={userDetails?.links} />
 
-        <UserDashStorage storageDetails={userDetails?.storage}/>
+        <UserDashStorage storageDetails={userDetails?.storage} />
       </div>
 
-      <div className='flex gap-10 h-[45%]'>
-        <UserDashFavGroups/>
+      <div className="flex gap-10 h-[45%]">
+        <UserDashFavGroups />
 
-        <UserDashCritLinks critLinks={userDetails?.critical_links}/>
+        <UserDashCritLinks critLinks={userDetails?.critical_links} />
       </div>
     </div>
-  )
-}
+  );
+};
 
-export default UserDashboard
+export default UserDashboard;
