@@ -1,21 +1,25 @@
-export default async function editUser(username: string, action: string, extra_storage?: number) {
+export default async function editUser(
+  username: string,
+  action: string,
+  extra_storage?: number
+) {
   const token = process.env.NEXT_PUBLIC_TEST_TOKEN;
   if (token) {
-    let data: { username: string; action: string; extra_storage?: number } = {
-      "username": username,
-      "action": action
+    const data: { username: string; action: string; extra_storage?: number } = {
+      username: username,
+      action: action,
     };
 
     if (action === 'add_extra') {
-      data["extra_storage"] = extra_storage;
+      data['extra_storage'] = extra_storage;
     }
 
     const res = await fetch(
       `https://api.immunefiles.com/api/api/auth/admin/user/edit?tenant=${
-        window.location.hostname.split(".")[0]
+        window.location.hostname.split('.')[0]
       }`,
       {
-        method: "PATCH",
+        method: 'PATCH',
         headers: {
           'Content-Type': 'application/json',
           Authorization: `Bearer ${token}`,
@@ -31,6 +35,6 @@ export default async function editUser(username: string, action: string, extra_s
       throw new Error('Error editing user');
     }
   } else {
-    throw new Error("No token found");
+    throw new Error('No token found');
   }
 }
